+++ src/config.rs
@@ -2,10 +2,10 @@
 //!
 //! [`clap`]: https://github.com/clap-rs/clap/
 
-/// The configuration parameters for the application.
+/// Configuration parameters for the application.
 #[derive(clap::Parser)]
 pub struct Config {
-    /// The connection URL for the Postgres database this application should use.
+    /// Connection URL for the Postgres database this application should use.
     #[clap(long, env)]
     pub database_url: String,
 }