+++ src/routes.rs
@@ -2,17 +2,18 @@
 //!
 //! [`axum`]: https://github.com/tokio-rs/axum
 
+use self::todo_item::{
+    create_todo_item::create_todo_item, delete_todo_item::delete_todo_item,
+    get_todo_item::get_todo_item, update_todo_item::update_todo_item,
+};
 use axum::{
     routing::{delete, get, patch, post},
     Router,
 };
 
-use self::todo_item::{
-    create_todo_item::create_todo_item, delete_todo_item::delete_todo_item,
-    get_todo_item::get_todo_item, update_todo_item::update_todo_item,
-};
+pub mod extractors;
+pub mod response_builders;
 
-pub mod response;
 pub mod todo_item {
     pub mod create_todo_item;
     pub mod delete_todo_item;