+++ src/routes/response_builders.rs
@@ -1,29 +1,22 @@
 //! Contains common axum responses for our routes
 
-use std::collections::HashMap;
-
 use axum::{http::StatusCode, Json};
-use serde_json::json;
+use serde_json::{json, Value};
+use std::collections::HashMap;
 use tracing::error;
 
-//TODO: transform in to a macro?
-pub fn json_error(code: String, message: String) -> (StatusCode, String) {
-    let body = Json(json!({ "code": code, "message": message }));
-    (StatusCode::NOT_FOUND, body.to_string())
-}
-
-/// Utility function for mapping any error into a `500 Internal Server Error`
-/// response.
-pub fn internal_error<E>(err: E) -> (StatusCode, String)
+/// Utility function for mapping any error into a `500 Internal Server Error` response.
+pub fn internal_error<E>(err: E) -> (StatusCode, axum::Json<Value>)
 where
     E: std::error::Error,
 {
     error!("internal_error: {}", err);
-    let body = Json(json!({ "code": "INTERNAL_SERVER_ERROR", "message": err.to_string() }));
-    (StatusCode::INTERNAL_SERVER_ERROR, body.to_string())
+    let body = json!({ "code": "INTERNAL_SERVER_ERROR", "message": err.to_string() });
+    (StatusCode::INTERNAL_SERVER_ERROR, Json(body))
 }
 
-pub fn validation_error(err: validator::ValidationErrors) -> (StatusCode, String) {
+/// Utility function for mapping an validator::ValidationErrors Error to an axum validation error http response
+pub fn validation_error(err: validator::ValidationErrors) -> (StatusCode, axum::Json<Value>) {
     let error_map: HashMap<String, Vec<String>> = err
         .field_errors()
         .into_iter()
@@ -38,11 +31,11 @@ pub fn validation_error(err: validator::ValidationErrors) -> (StatusCode, String
         })
         .collect();
 
-    let body = Json(json!({
+    let body = json!({
         "code": "VALIDATION_ERROR",
         "message": "Validation error occurred",
         "errors": error_map
-    }));
+    });
 
-    (StatusCode::BAD_REQUEST, body.to_string())
+    (StatusCode::BAD_REQUEST, Json(body))
 }