+++ src/routes/todo_item/delete_todo_item.rs
@@ -1,9 +1,9 @@
+use crate::routes::response_builders::internal_error;
 use axum::{extract::Path, http::StatusCode, Json};
 use serde::Serialize;
+use serde_json::{json, Value};
 use sqlx::{postgres::PgQueryResult, PgPool};
 use uuid::Uuid;
-
-use crate::routes::response::{internal_error, json_error};
 
 #[derive(Serialize)]
 pub struct DeleteTodoItemResponse {
@@ -13,7 +13,7 @@ pub struct DeleteTodoItemResponse {
 pub async fn delete_todo_item(
     db: axum::Extension<PgPool>,
     Path(todo_item_id): Path<Uuid>,
-) -> Result<Json<DeleteTodoItemResponse>, (StatusCode, String)> {
+) -> Result<Json<DeleteTodoItemResponse>, (StatusCode, axum::Json<Value>)> {
     let db_result: Result<PgQueryResult, sqlx::Error> = sqlx::query!(
         r#"
             DELETE
@@ -28,10 +28,8 @@ pub async fn delete_todo_item(
     let db_result = db_result.map_err(internal_error)?;
 
     if db_result.rows_affected() == 0 {
-        Err(json_error(
-            "NOT_FOUND".to_string(),
-            format!("Todo Item with id '{}' not found", todo_item_id),
-        ))
+        let body = json!({ "code": "NOT_FOUND".to_string(), "message": format!("Todo Item with id '{}' not found", todo_item_id) });
+        Err((StatusCode::NOT_FOUND, Json(body)))
     } else {
         Ok(Json(DeleteTodoItemResponse { success: true }))
     }