+++ src/routes/todo_item/get_todo_item.rs
@@ -1,14 +1,15 @@
+use crate::domain::entities::todo_item::TodoItem;
 use crate::domain::value_objects::priority_level::PriorityLevel;
-use crate::routes::response::internal_error;
-use crate::{domain::entities::todo_item::TodoItem, routes::response::json_error};
+use crate::routes::response_builders::internal_error;
 use axum::{extract::Path, http::StatusCode, Json};
+use serde_json::{json, Value};
 use sqlx::PgPool;
 use uuid::Uuid;
 
 pub async fn get_todo_item(
     db: axum::Extension<PgPool>,
     Path(todo_item_id): Path<Uuid>,
-) -> Result<Json<TodoItem>, (StatusCode, String)> {
+) -> Result<Json<TodoItem>, (StatusCode, axum::Json<Value>)> {
     let db_result: Result<TodoItem, sqlx::Error> = sqlx::query_as!(
         TodoItem,
         r#"
@@ -22,10 +23,10 @@ pub async fn get_todo_item(
     .await;
 
     let todo_item = db_result.map_err(|e| match e {
-        sqlx::Error::RowNotFound => json_error(
-            "NOT_FOUND".to_string(),
-            format!("Todo Item with id '{}' not found", todo_item_id),
-        ),
+        sqlx::Error::RowNotFound => {
+            let body = json!({ "code": "NOT_FOUND".to_string(), "message": format!("Todo Item with id '{}' not found", todo_item_id) });
+            (StatusCode::NOT_FOUND, Json(body))
+        },
         e => internal_error(e),
     })?;